requirements.txt
pre_commit==4.5.0
fastapi==0.116.1
flake8==7.0.0
python-dotenv==1.2.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.14
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.6.1
sqlmodel==0.0.27
asyncpg==0.27.0